+++ src/transaction/placeOrder.ts
@@ -83,8 +83,6 @@ export interface IObject {
      * WAITER許可証
      */
     passport?: waiter.passport.IPassport;
-    // 使用していないので廃止(2023-08-15~)
-    // authorizeActions?: IAuthorizeAction<IAuthorizeActionAttributes<any, any>>[];
     potentialActions?: {
         givePointAward?: IGivePointAwardParams[];
     };
@@ -188,7 +186,7 @@ export interface IConfirmParams {
     result?: IResultParams;
 }
 export type IOrderAsResult = OrderFactory.IOrder & {
-    acceptedOffers: OrderFactory.IAcceptedOffer<OrderFactory.IItemOffered>[];
+    // acceptedOffers: OrderFactory.IAcceptedOffer<OrderFactory.IItemOffered>[]; // 廃止(2024-02-04~)
 };
 export interface IAuthorizeActionAsResult { id: string; }
 /**
@@ -199,6 +197,10 @@ export interface IResult {
      * 注文
      */
     order: IOrderAsResult;
+    /**
+     * 取引確定時の同期的な注文コード発行に対応(2024-02-05~)
+     */
+    code?: string;
     /**
      * 承認アクションID(2024-01-17~)
      */